Update README.md
# 6_Digit_7_segment_display_clock
This project is a digital clock that displays time in HH:MM:SS format using a 6-digit 7-segment display controlled by an Arduino. Time is tracked using the millis() function, and users can manually set hours and minutes with push buttons. The display brightness can be adjusted, and the clock supports both 12-hour and 24-hour modes.